wrapper_server/requirements.txt
# Wrapper Server Dependencies
# FastAPI web framework
fastapi>=0.109.0
uvicorn>=0.27.0

# HTTP client for OpenCode API calls
httpx>=0.25.0

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data validation
pydantic>=2.5.0

# Logging
loguru>=0.7.0

# File uploads
python-multipart>=0.0.6

# Process management (for opencode-manager.py)
psutil>=5.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.25.0  # For testing async endpoints
